+++ lib/minting/mint.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
 require 'minting/mint/currency'
-require 'minting/mint/refinements'
+require 'minting/mint/currency_store'
 require 'minting/mint/registry'
+require 'minting/mint/refinements'